[FEAT] 상품 도메인 이벤트 처리 초기 버전 추가  (#19)
* feat: 상품 등록과 타임딜 설정 분리
- 상품 등록과 타임딜 설정을 분리하여 mvp에선 타임딜 등록 요청 시 요청 시각부터 설정 시각까지의 타임딜이 바로 실행될 수 있도록 구현

* test: 변경된 타임딜 설정 시기에 따라 테스트코드 수정 후 확인
- 테스트 중 확인된 도메인 로직 오류 수정

* refactor: 타임딜 설정 시 localDateTime 여러 번 호출하는 부분 수정

* feat: 컨트롤러 프레임 및 표현 계층 dto 작성
- 외부 컨트롤러 프레임 구현
- 표현 계층의 dto 작성
- TimeDealSchedule의 생성자 access 속성 변경

* refactor: 코드래빗 피드백 반영
- 컨트롤러 내 중복 매핑 수정, 경로변수 사용 부분 수정
- dto 내에서 price 검증 추가, 타입, 필드명 일치

* feat: 응용계층 dto 구현 및 응용계층 구현을 위해 필요한 설정 추가
- JpaAuditConfig 구현, EnableJpaAuditing 해당 위치에 사용
- ProductMapper 구현
- UserContext 및 관련 설정 파일 추가

* Update src/main/java/org/pgsg/product/config/jpa/JpaAuditConfig.java


* Update src/main/java/org/pgsg/product/config/security/UserContextInterceptor.java


* Update src/main/java/org/pgsg/product/config/web/WebConfig.java


* Add validation to updateTimeDealSchedule method

* feat: 코드래빗 피드백 반영
- 유저 컨텍스트 주입 방식을 기존 헤더 추출에서 SecurityContextHolder 추출로 변경해서 보안성 강화

* feat: 상품 추가, 삭제 구현 및 커스텀 예외 처리 final화
- 상품 추가, 삭제 구현
- 커스텀 예외 처리를 문자열 직접 입력에서 상수 사용식으로 변경

* test: 상품 추가, 삭제 기능 및 dto 테스트
- infra계층의 repository에 override 추가

* Refactor deleteProduct to ensure userId is not null

* feat: 상품 수정 부분 추가
- 상품 정보 수정, 타임딜 설정, 판매 취소 구현

* test: 상품 수정 관련 테스트 진행
- 상품수정, 타임딜 설정, 판매 취소 테스트
- 테스트 중 필요 부분 수정

* Update src/main/java/org/pgsg/product/application/service/ProductCommandService.java


* feat: 상품 조회 부분 구현
- ProductQueryService 구현
- 병합 중 누락된 부분 복구
- mapper에 엔티티->dto 변환 추가 및 적용

* fix: 병합 중 누락 코드 복구

* refactor: 계층별 매퍼 분리

* refactor: 필드명이 다른 속성의 매핑을 위해 애너테이션 추가

* refactor: 예외 처리 방식 변경
- 예외 처리 방식을 기존 예외 class 내 상수 -> enum 사용으로 변경
- yml파일 내 예외명 또한 추후 계층형으로 전환을 위해 도트 표기법으로 수정
- application.yml 내 예외 import 삭제
- 병합 중 누락 부분 복구

* refactor: 예외 처리 방식 변경
- 예외 처리 방식을 기존 예외 class 내 상수 -> enum 사용으로 변경
- yml파일 내 예외명 또한 추후 계층형으로 전환을 위해 도트 표기법으로 수정
- application.yml 내 예외 import 삭제
- 병합 중 누락 부분 복구

* test: 예외 검증 테스트 진행
- 빌드 중 발견된 부분 수정

* feat: 이벤트 연결
- 타임딜 설정 시 outbox에 이벤트 저장
- 예약/거래 진행과정에 따라 inbox에 이벤트 수신

* feat: 변경된 common 이벤트 처리 반영 등
- domainType, eventType 직접 주입하도록 수정
- configs의 토픽명을 사용할 수 있도록topicConfig 작성 및 적용

* refactor: 실행 중 발견사항 수정
- TopicConfig: Reservation 선언 누락 부분 추가
- configs/common/application-topics.yml 파일 인식 불가로 일단 하드코딩 테스트 후 재시도 예정
-config-server 주소 매핑 오류로 yml에 직접 추가 및 yml 파일 명 bootstrap.yml로 변경
-JPAAuditConfig: AuditorAware이 common 모듈에서도 선언되어 중복 선언으로 파일 삭제

* refactor: 실행 중 변경 사항 - 이벤트 발행 테스트 성공
- 조회 api 응답 dto들에 selleId와 status 추가, mapper에도 해당 부분에 대한 mapping 추가
- event 발행 방식을 kafkaTemplate에서 ApplicationEventPublisher로 변경

* refactor: 이벤트 구독 처리 테스트 중 변경사항
- db: h2->postgres 변경
- ProductApplication: @EntityScan 추가
- Product: 스케줄 등록 시 바로 예약으로 넘어가므로 PENDING_RESERVATION에서 바로 RESERVING으로 넘어가도록 수정

* chore: 로컬에서 이벤트 테스트를 위해 설정했던 부분 해제
 - settings.gradle: common 모듈 로컬 연결 해제
 - build.gradle: common 모듈 버전업

* Update .gitignore


* refactor: 코드래빗 피드백 반영
- build.gradle: spring-cloud-starter-config 추가
- TopicConfig: 선언한 필드에 대해서 검증 추가
- ProductKafkaConsumer: id추출 부분을 공통으로 분리하고, if문으로 명시적 검증 진행

* refactor: 코드래빗 피드백 반영
- build.gradle: spring-cloud-starter-config 추가
- TopicConfig: 선언한 필드에 대해서 검증 추가
- ProductKafkaConsumer: id추출 부분을 공통으로 분리하고, if문으로 명시적 검증 진행

* refactor: 이벤트 발행 시 직렬화 부분 수정
- ProductCommandService: 이벤트 발행 시 payload를 직렬화해서 넣을 경우 outboxService에서 직렬화가 한번 더 일어나므로 발행 부분에선 직렬화 제외

* chore: docker-compose.yaml 생성

* refactor: 이벤트에서 상품id 추출 중 예외 발생 시 작성되는 로그 내용에서 value 제외
- common 모듈 0.2.7 버전으로 갱신

* refactor: 실행 중 변동 사항 적용

* refactor: 로그 기록 변경
- bootstrap.yml: logging 레벨 warn으로 올림
- ProductCommandService: 각 단계 별 로그 추가

* feat: 인증 관련 부분 추가
- ProductSecurityConfig: 상품 내 인가 관련 설정 추가, 인증 유저 정보는 LoginFilter에 의해 SecurityContext에 저장됨
- 기존에 사용하던 UserContext 및 UserContextInterceptor 삭제
- ProductCommandService: 수정 및 삭제 시 유저 권한의 사용자의 id 검증 추가, 관련된 401,403 예외 yml및 enum에 추가

* feat: 인가 및 이벤트 반영 후 테스트 코드 수정
- WebConfig 삭제
- ProductCommandServiceTest: 인가 및 이벤트 부분 추가

* chore: 항상 common 모듈의 최신 버전을 사용하도록 설정

* test: 도메인 테스트 오류 부분 수정
- 테스트 객체에 상태 주입 추가

* test: 도메인 테스트 오류 부분 수정
- 테스트 객체에 상태 주입 추가
- 상품측 커스텀 보안설정이 불필요하므로 삭제

---------

+++ src/main/java/org/pgsg/product/application/service/ProductCommandService.java
@@ -3,21 +3,31 @@ package org.pgsg.product.application.service;
 import static org.pgsg.product.global.exception.ProductErrorCode.*;
 
 import java.time.LocalDateTime;
+import java.util.Objects;
 import java.util.UUID;
 
+import org.pgsg.common.event.OutboxEvent;
 import org.pgsg.common.exception.CustomException;
+import org.pgsg.common.util.SecurityUtil;
+import org.pgsg.config.security.UserDetailsImpl;
 import org.pgsg.product.application.dto.command.CreateProductCommand;
 import org.pgsg.product.application.dto.command.UpdateProductCommand;
 import org.pgsg.product.application.dto.command.UpdateTimeDealCommand;
 import org.pgsg.product.application.dto.result.CreateProductResult;
 import org.pgsg.product.application.dto.result.UpdateProductResult;
-import org.pgsg.product.domain.model.TimeDealSchedule;
+import org.pgsg.product.application.mapper.ProductApplicationMapper;
+import org.pgsg.product.domain.event.ProductCreatedEvent;
 import org.pgsg.product.domain.model.Product;
+import org.pgsg.product.domain.model.TimeDealSchedule;
 import org.pgsg.product.domain.repository.ProductRepository;
+import org.pgsg.product.global.config.TopicConfig;
+import org.pgsg.product.global.exception.ProductErrorCode;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 /*
 * 상품 추가, 삭제, 수정 조회, 이벤트 연결
@@ -25,28 +35,35 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 @Transactional
+@Slf4j
 public class ProductCommandService {
 	private final ProductRepository productRepository;
+	private final ProductApplicationMapper mapper;
+	private final TopicConfig topicConfig;
+	private final ApplicationEventPublisher eventPublisher;
 
 	public CreateProductResult createProduct(CreateProductCommand command) {
 		Product product =Product.create(
 			command.name(),command.price(),command.description());	//todo: 스케줄 입력 시기 변경 후 수정 필요
 
 		Product saved=productRepository.save(product);
-
-		return new CreateProductResult(saved.getName(),saved.getPrice(),saved.getDescription());
+		log.info("Created product: productId:{}, userId:{}", saved.getId(),saved.getCreatedBy());
+		//todo: 스케줄링 도구 적용 시 생성 후 스케줄이 등록되도록 고도화 예정
+		return new CreateProductResult(saved.getId(),saved.getName(),saved.getPrice(),saved.getDescription());
 	}
 
 	public void deleteProduct(UUID id) {
 		Product product=findById(id);
+		checkAuthorization(product.getCreatedBy());
 
-		UUID userId = /*Objects.requireNonNull(UserContext.getUserId(), "인증 사용자 정보가 없습니다.");*/
-			UUID.randomUUID();	//todo: 로컬 테스트용, 인증 서비스 연결 후 수정
+		UUID userId = SecurityUtil.getCurrentUserIdOrThrow();
 		product.deleteProduct(userId);
+		log.info("Deleted product. productId:{}, userId:{}", id, userId);
 	}
 
 	public UpdateProductResult updateProduct(UUID id, UpdateProductCommand command) {
 		Product product = findById(id);
+		checkAuthorization(product.getCreatedBy());
 
 		//todo: timeDealSchedule 설정부분 리팩토링 후 수정 예정
 		TimeDealSchedule newSchedule = command.endTime() == null ? null
@@ -56,6 +73,7 @@ public class ProductCommandService {
 		product.update(command.name(), command.price(), command.description(), newSchedule);
 
 		Product saved = productRepository.saveAndFlush(product);
+		log.info("Updated product: productId:{}, userId:{}", saved.getId(), saved.getModifiedBy());
 
 		TimeDealSchedule schedule = saved.getTimeDealSchedule();
 		return new UpdateProductResult(
@@ -68,29 +86,69 @@ public class ProductCommandService {
 	}
 	public UpdateProductResult setTimeDeal(UUID id, UpdateTimeDealCommand command) {
 		Product product = findById(id);
+		checkAuthorization(product.getCreatedBy());
 
 		product.setTimeDealSchedule(command.endTime());
 
 		Product saved=productRepository.saveAndFlush(product);
+		log.info("Set Timedeal: productId:{}, userId:{}", saved.getId(), saved.getModifiedBy());
 
-		//todo: 타임딜 설정 후 이벤트 발행 부분 추가 예정 - mvp 이후 이벤트 발행 위치 변경 예정
+		//todo: mvp 이후 이벤트 발행 위치 변경 예정
+		ProductCreatedEvent payload = mapper.toCreatedEvent(product);
+		String eventType=topicConfig.getProduct().getCreated();
+		OutboxEvent event=new OutboxEvent(saved.getId(),  saved.getId(),"Product", eventType, payload);
+
+		eventPublisher.publishEvent(event);
+		log.info("event is published. productId:{}, eventType:{}", saved.getId(), eventType);
+
 
 		return new UpdateProductResult(saved.getName(), saved.getPrice(), saved.getDescription(),
 			saved.getTimeDealSchedule().getStartTime(), saved.getTimeDealSchedule().getEndTime());
 	}
 
-	public void cancelSaleProduct(UUID id) {
+	public void cancelSale(UUID id) {
 		Product product = findById(id);
 		product.cancelSale();
+		log.info("Sale is cancelled. productId:{}", product.getId());
 
 		// UUID userId = /*Objects.requireNonNull(UserContext.getUserId(), "인증 사용자 정보가 없습니다.");*/
 		// 	UUID.fromString("00000000-0000-0000-0000-000000000000");	//todo: 로컬 테스트용, 인증 서비스 연결 후 수정
 		// product.deleteProduct(id);	//todo: 삭제와 판매 취소를 동일하게 할지 좀 더 고려
 	}
 
+	public void completeTrade(UUID id) {
+		Product product = findById(id);
+		product.complete();
+		log.info("Trade is completed. productId:{}", product.getId());
+	}
 
-	private Product findById(UUID id) {
-		return productRepository.findById(id)
+	public void completeReservation(UUID id) {
+		Product product = findById(id);
+		product.startTrade();
+		log.info("Reservation is completed. productId:{}", product.getId());
+	}
+
+	public void pendingSale(UUID id) {
+		Product product = findById(id);
+		product.revertToReserving();
+		log.info("This product revert to reserving. productId:{}", product.getId());
+	}
+
+
+	private Product findById(UUID productId) {
+		return productRepository.findById(productId)
 			.orElseThrow(()->new CustomException(ProductNotFoundException));
+	}
+
+	private void checkAuthorization(UUID sellerId) {
+		String userRole=SecurityUtil.getCurrentUser()
+			.map(UserDetailsImpl::getUserRole)
+			.orElse(null);
+		if(userRole==null)
+			throw new CustomException(Forbidden);
+		else if(userRole.equals("USER")) {
+			if(!SecurityUtil.getCurrentUserIdOrThrow().equals(sellerId))
+				throw new CustomException(Unauthorized);
+		}
 	}
 }
